+++ src/LBMSCommands/BorrowBook.java
@@ -12,7 +12,7 @@ import Library.Library;
 public class BorrowBook implements LBMSCommand
 {
     private Library library;
-    private Integer visitorID;  // This may have to be a string that is converted to int in execute...
+    private Long visitorID;  // This may have to be a string that is converted to int in execute...
     private ArrayList<String> bookID;
 
     /**
@@ -22,7 +22,7 @@ public class BorrowBook implements LBMSCommand
      * @param visitorID - The unique 10-digit ID of the visitor.
      * @param bookID - The comma-separated list of IDs for the books to be borrowed by the visitor. 5 at most.
      */
-    public BorrowBook(Library library, Integer visitorID, ArrayList<String> bookID)
+    public BorrowBook(Library library, Long visitorID, ArrayList<String> bookID)
     {
         this.library = library;
         this.visitorID = visitorID;
