+++ src/LBMSCommands/EndVisit.java
@@ -11,7 +11,7 @@ import Library.Library;
 public class EndVisit implements LBMSCommand
 {
     private Library library;
-    private Integer visitorID; // This may have to be a string that is converted to int...
+    private Long visitorID; // This may have to be a string that is converted to int...
 
     /**
      * Constructs an EndVisit command to be executed.
@@ -19,7 +19,7 @@ public class EndVisit implements LBMSCommand
      * @param library - the library that the visitor was visiting.
      * @param visitorID - the ID of an active visitor.
      */
-    public EndVisit(Library library, Integer visitorID)
+    public EndVisit(Library library, Long visitorID)
     {
         this.library = library;
         this.visitorID = visitorID;
